move `OpeningStatistic` to the new `types` module
