+++ berserk/types/__init__.py
@@ -3,7 +3,9 @@ from __future__ import annotations
 import requests
 
 from .team import Team
+from .opening_explorer import OpeningStatistic
 
 __all__ = [
     "Team",
+    "OpeningStatistic"
 ]
